Extract playground auth session hook
